Use CSI driver to determine unique name for migrated in-tree plugins
processVolumeAttachments currently only tries to find PVs with
uncertain attachemnt state. This will however lead to false postives for
CSI migrated PVs, as the unique volume does not match between the original
and migrated PV. This commit falls back to using the unique name of the
CSI PV when migration for the in-tree plugin is enabled.

The described false positives are an issue because later the attach/detach
controller tries to detach the volume as it thinks the volume should not be
attached to the node, ignoring that the CSI driver still thinks that the
same volme must be (and is) attached to this node. This causes POD mounted
volumes to loose the backing EBS storage and then running into all kinds of
follow up errors (io related).